+++ defaults.js
@@ -11,6 +11,7 @@ const defaults = {
   scripts: {},
   skip: {},
   dryRun: false,
+  tagForce: false,
   gitTagFallback: true,
   preset: require.resolve('conventional-changelog-conventionalcommits')
 }
@@ -18,7 +19,7 @@ const defaults = {
 /**
  * Merge in defaults provided by the spec
  */
-Object.keys(spec.properties).forEach(propertyKey => {
+Object.keys(spec.properties).forEach((propertyKey) => {
   const property = spec.properties[propertyKey]
   defaults[propertyKey] = property.default
 })
@@ -27,13 +28,10 @@ Object.keys(spec.properties).forEach(propertyKey => {
  * Sets the default for `header` (provided by the spec) for backwards
  * compatibility. This should be removed in the next major version.
  */
-defaults.header = '# Changelog\n\nAll notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.\n'
+defaults.header =
+  '# Changelog\n\nAll notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.\n'
 
-defaults.packageFiles = [
-  'package.json',
-  'bower.json',
-  'manifest.json'
-]
+defaults.packageFiles = ['package.json', 'bower.json', 'manifest.json']
 
 defaults.bumpFiles = defaults.packageFiles.concat([
   'package-lock.json',